rename redux store directory (don't do that again)
