+++ client/src/App.tsx
@@ -12,8 +12,8 @@ import React, { ReactElement } from 'react';
 
 import { Button, Container } from 'react-bootstrap';
 import { Route, Routes, useNavigate } from 'react-router-dom';
-import { useAppSelector } from 'redux/hooks';
-import { RootState } from 'redux/store';
+import { useAppSelector } from 'store/hooks';
+import { RootState } from 'store/store';
 import './App.scss';
 
 function App(): ReactElement {
@@ -69,9 +69,7 @@ function App(): ReactElement {
                   <HtCard>
                     <HtCard.Header title="This is not the page you're looking for..." />
                     <HtCard.Body className="d-grid justify-content-center">
-                      <h1 className="display-1 d-flex justify-content-center">
-                        404
-                      </h1>
+                      <h1 className="display-1 d-flex justify-content-center">404</h1>
                       <h4>Resource not found</h4>
                     </HtCard.Body>
                     <HtCard.Footer>
