+++ client/src/components/Nav/Sidebar.tsx
@@ -2,7 +2,7 @@ import React, { ReactElement, useState } from 'react';
 import { Link } from 'react-router-dom';
 import { useSelector } from 'react-redux';
 import { Button, Collapse } from 'react-bootstrap';
-import { RootState } from 'redux/store';
+import { RootState } from 'store/store';
 
 /**
  * Vertical sidebar for navigation that disappears when the viewport is small
