+++ client/src/features/login/Login.tsx
@@ -2,8 +2,8 @@ import React, { ReactElement, useEffect } from 'react';
 import { useForm } from 'react-hook-form';
 import { yupResolver } from '@hookform/resolvers/yup';
 import * as Yup from 'yup';
-import { login } from 'redux/store';
-import { useAppDispatch, useAppSelector } from 'redux/hooks';
+import { login } from 'store/store';
+import { useAppDispatch, useAppSelector } from 'store/hooks';
 import { useNavigate } from 'react-router-dom';
 import { Form } from 'react-bootstrap';
 import useTitle from '../../hooks/useTitle';
