+++ client/src/index.tsx
@@ -3,7 +3,7 @@ import { createRoot } from 'react-dom/client';
 import { Provider } from 'react-redux';
 import { BrowserRouter } from 'react-router-dom';
 import App from './App';
-import { store } from 'redux/store';
+import { store } from 'store/store';
 
 // initiate the mock service worker if deployed in TEST environment
 // intercepts API calls and returns fake/test responses
